refactor: Bookmark Category, OpenType vo 패키지로 이동
