+++ src/main/java/com/meoguri/linkocean/controller/bookmark/CategoryController.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.meoguri.linkocean.controller.common.SliceResponse;
-import com.meoguri.linkocean.domain.bookmark.entity.Bookmark;
+import com.meoguri.linkocean.domain.bookmark.entity.vo.Category;
 
 @RequestMapping("/api/v1/categories")
 @RestController
@@ -14,6 +14,6 @@ public class CategoryController {
 	@GetMapping
 	public SliceResponse<String> getAllCategories() {
 
-		return SliceResponse.of("categories", Bookmark.Category.getKoreanNames());
+		return SliceResponse.of("categories", Category.getKoreanNames());
 	}
 }
