+++ src/main/java/com/meoguri/linkocean/domain/bookmark/entity/vo/OpenType.java
@@ -1,0 +1,23 @@
+package com.meoguri.linkocean.domain.bookmark.entity.vo;
+
+/**
+ * 북마크의 공개 범위
+ */
+public enum OpenType {
+	/* 전체공개 */
+	ALL,
+
+	/* 팔로워 대상 공개 */
+	PARTIAL,
+
+	/* 개인 공개 */
+	PRIVATE;
+
+	public String getName() {
+		return name().toLowerCase();
+	}
+
+	public static OpenType of(String arg) {
+		return OpenType.valueOf(arg.toUpperCase());
+	}
+}
